extract sample data to separate file
